Conditionally install bcfips jars when building/testing observabilitySRE (#17359)
* Conditionally install bcfips jars when building for observabilitySRE

This commit implements a pattern for performing specific gradle tasks based on a
newly named "fedrampHighMode" option. This option is used to configure tests to
run with additional configuration specific to the observabilitySRE use case.
Similarly the additional jar dependencies for bouncycastle fips providers are
conditionally installed gated on the "fedrampHighMode" option.

In order to ensure the the "fedrampHighMode" option persists through the layers
of sub-processes spawned between gradle and rake we store and respect an
environment variable FEDRAMP_HIGH_MODE. This may be useful generally in building
the docker image.

Try codereview suggestion

* Use gradle pattern for setting properties with env vars

Gradle has a mechanism for setting properties with environment variables
prefixed with `ORG_GRADLE_PROJECT`. This commit updates the gradle tasks to use
that pattern.

See
https://docs.gradle.org/current/userguide/build_environment.html#setting_a_project_property
for details.